+++ src/main/kotlin/coder/format/parse/png/chunk/PNGPrintOffsetChunk.kt
@@ -4,8 +4,8 @@ import org.bread_experts_group.coder.format.parse.png.PNGPrintOffsetUnit
 import java.nio.channels.SeekableByteChannel
 
 class PNGPrintOffsetChunk(
-	val x: Int,
-	val y: Int,
+	val x: Long,
+	val y: Long,
 	val unit: PNGPrintOffsetUnit,
 	window: SeekableByteChannel
 ) : PNGChunk("oFFs", window) {
